requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
pydantic
biopython
numpy
pandas
scikit-learn
matplotlib
seaborn
scipy
supabase
pydantic-settings